% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fertility_rates.R
\name{calculate_fertility_rates}
\alias{calculate_fertility_rates}
\title{calcualtes age specific and total fertility rates per year}
\usage{
calculate_fertility_rates(data)
}
\arguments{
\item{data}{births data frame}
}
\value{
tablulated age specific and total fertility rates per year
}
\description{
calcualtes age specific and total fertility rates per year
}
\examples{
fertility_rates <- calculate_fertility_rates(bth_data)

}
