% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_3_1.R
\name{create_t3.1}
\alias{create_t3.1}
\title{Create Table 3.1 and output csv file}
\usage{
create_t3.1(births_data, deaths_data)
}
\arguments{
\item{births_data}{name of births data frame}

\item{deaths_data}{name of deaths data frame}
}
\value{
Data frame with tabulated result
}
\description{
Create Table 3.1 and output csv file
}
\examples{
t3.1 <- create_t3.1(bth_data, dth_data)

}
