% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_3_10.R
\name{create_t3.10}
\alias{create_t3.10}
\title{Create Table 3.10}
\usage{
create_t3.10(data, date_var, data_year, tablename = NA)
}
\arguments{
\item{data}{dataframe being used}

\item{date_var}{occurrence data being used e.g. dobyr, dodyr etc}

\item{data_year}{year the data is for}

\item{tablename}{name of the table being saved as a csv file}
}
\value{
data frame of tabulated results
}
\description{
Create Table 3.10
}
\examples{
t3.10 <- create_t3.10(dth_data, date_var = dodyr, data_year = 2022, tablename = "Table_3_10")

}
