% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_3_11.R
\name{create_t3.11_and_3.12}
\alias{create_t3.11_and_3.12}
\title{Create Tables 3.11 or 3.12}
\usage{
create_t3.11_and_3.12(
  data,
  date_var,
  data_year,
  by_var,
  tablename = NA,
  topic = NA
)
}
\arguments{
\item{data}{data frame being used}

\item{date_var}{the occurrence year being used e.g. dobyr or dodyr}

\item{data_year}{the year of the data}

\item{by_var}{what variable the data is being redistributed by}

\item{tablename}{name of the table being saved as a csv file}

\item{topic}{whether the table is births or deaths data}
}
\value{
data frames for tabulated versions of Table 3.4 and 3.6
}
\description{
Create Tables 3.11 or 3.12
}
\examples{
t3.11 <- create_t3.11_and_3.12(bth_data, dobyr, data_year = 2022, by_var = fert_age_grp, tablename = "Table_3_11", topic = "births")
t3.12 <- create_t3.11_and_3.12(dth_data, dodyr, data_year = 2022, by_var = age_grp_80, tablename = "Table_3_12", topic = "deaths")

}
