% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_3_2_&_3_3.R
\name{create_t3.2_t3.3}
\alias{create_t3.2_t3.3}
\title{Creates Tables 3.2 & 3.3 and outputs csv files}
\usage{
create_t3.2_t3.3(data, occ_var, topic = NA, tablename = "table_3_2")
}
\arguments{
\item{data}{name of data frame being used}

\item{occ_var}{year variable for births or deaths occurrences}

\item{topic}{whether births or deahts is being run}

\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated results
}
\description{
Creates Tables 3.2 & 3.3 and outputs csv files
}
\examples{
t3.2 <- create_t3.2_t3.3(bth_data, occ_var = dobyr, topic = "births", tablename = "Table_3_2")
}
