% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_3_8_&_3_9.R
\name{create_t3.8_and_t3.9}
\alias{create_t3.8_and_t3.9}
\title{Creates Tables 3.8 and 3.9}
\usage{
create_t3.8_and_t3.9(
  data,
  est_data,
  date_var,
  data_year,
  by_var,
  topic = NA,
  tablename = NA
)
}
\arguments{
\item{data}{name of data frame being used}

\item{est_data}{data frame of estimated data being used}

\item{date_var}{occurrence data variable being used}

\item{data_year}{year of data}

\item{by_var}{the occurrence year being used e.g. dobyr or dodyr}

\item{topic}{whether births or deahts is being run}

\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated results
}
\description{
Creates Tables 3.8 and 3.9
}
\examples{
t3.8 <- create_t3.8_and_t3.9(data = bth_data, est_data = bth_est,
date_var = dobyr, data_year = 2022, by_var = rgn,
topic = "births", tablename = "Table_3_8")
t3.9 <- create_t3.8_and_t3.9(data = dth_data, est_data = dth_est,
                            date_var = dodyr, data_year = 2022, by_var = rgn,
                            topic = "deaths", tablename = "Table_3_9")

}
