% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_3_8_&_3_9.R
\name{create_t3.8_t3.9}
\alias{create_t3.8_t3.9}
\title{Creates Tables 3.8 and 3.9}
\usage{
create_t3.8_t3.9(data, est_data, by_var, topic = NA, tablename = NA)
}
\arguments{
\item{data}{name of data frame being used}

\item{est_data}{data frame of estimated data being used}

\item{by_var}{the occurrence year being used e.g. dobyr or dodyr}

\item{topic}{whether births or deahts is being run}

\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated results
}
\description{
Creates Tables 3.8 and 3.9
}
\examples{
t3.8 <- create_t3.8_t3.9(bth_data, bth_est, rgn, topic = "births", tablename = "Table_3_8")

}
