% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_4_2.R
\name{create_t4.2}
\alias{create_t4.2}
\title{Creates Table 4.2}
\usage{
create_t4.2(
  data,
  est_data,
  date_var,
  data_year = 2022,
  tablename = "Table_4_2"
)
}
\arguments{
\item{data}{data frame being used}

\item{est_data}{data frmae of estimated data being used}

\item{date_var}{name of date variable}

\item{data_year}{year the data is for}

\item{tablename}{name of the table being saved as a csv file}
}
\value{
data frame of tabulated counts
}
\description{
Creates Table 4.2
}
\examples{
t4.2 <- create_t4.2(data = bth_data, est_data = bth_est, date_var = dobyr,
                             data_year = 2022, tablename = "Table_4_2")
}
