% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_4_8.R
\name{create_t4.8}
\alias{create_t4.8}
\title{creates table 4.8}
\usage{
create_t4.8(
  data,
  est_data,
  date_var,
  data_year = 2022,
  by_var = NA,
  tablename = "Table_4_8"
)
}
\arguments{
\item{data}{data frame being used}

\item{est_data}{data frame for estimate data}

\item{date_var}{occurrence data variable being used}

\item{data_year}{year of data}

\item{by_var}{variable the data is grouped by}

\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated result
}
\description{
creates table 4.8
}
\examples{
t4.8 <- create_t4.8(bth_data, bth_est, dobyr, data_year = 2022, by_var = rgn, tablename = "Table_4_8")
}
