% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_5_1.R
\name{create_t5.1}
\alias{create_t5.1}
\title{Calculates Table 4.1 Births summary table}
\usage{
create_t5.1(data, tablename = "Table_5_1")
}
\arguments{
\item{data}{births data frame}

\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated results
}
\description{
Calculates Table 4.1 Births summary table
}
\examples{
t5.1 <- create_t5.1(dth_data, tablename = "Table_5_1")

}
