% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_5_11.R
\name{create_t5.11}
\alias{create_t5.11}
\title{creates Table 5.11}
\usage{
create_t5.11(data, date_var, tablename = NA)
}
\arguments{
\item{data}{data frame being used}

\item{date_var}{occurrence data variable being used}

\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated result
}
\description{
creates Table 5.11
}
\examples{
t5.11 <- create_t5.11(bth_data, date_var = dobyr, tablename = "Table_5_11")

}
