% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_5_12.R
\name{create_t5.12}
\alias{create_t5.12}
\title{Create Table 5.12}
\usage{
create_t5.12(data, date_var, data_year = 2022, tablename = NA)
}
\arguments{
\item{data}{data frame being used}

\item{date_var}{occurrence data being used e.g. dobyr, dodyr etc}

\item{data_year}{}

\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tablutated results
}
\description{
Create Table 5.12
}
\examples{
t5.12 <- create_t5.12(bth_data, date_var = dobyr, data_year = 2022, tablename = "Table_5_12")

}
