% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_5_3_&_5_4.R
\name{create_t5.3_and_t5.4}
\alias{create_t5.3_and_t5.4}
\title{Calculates Tables 5.3 & 5.4
Deaths by place of occurrence and place of usual residence of decedent}
\usage{
create_t5.3_and_t5.4(
  data,
  date_var,
  data_year = 2022,
  sex_filter = NA,
  tablename = "Table_5_3"
)
}
\arguments{
\item{data}{dataframe being used}

\item{date_var}{occurrence data being used e.g. dobyr, dodyr etc}

\item{data_year}{year the data is for}

\item{sex_filter}{male or female}

\item{tablename}{name of the table being saved as a csv file}
}
\value{
data frame of tabulated results
}
\description{
Calculates Tables 5.3 & 5.4
Deaths by place of occurrence and place of usual residence of decedent
}
\examples{
t5.3 <- create_t5.3_and_t5.4(dth_data, dodyr, 2022, sex_filter = "male", tablename = "Table_5_3")

}
