% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_5_5.R
\name{create_t5_5}
\alias{create_t5_5}
\title{Calculates Table 5.5 Deaths by place of occurrence and site of occurrence}
\usage{
create_t5_5(data, date_var = dodyr, data_year = 2022, tablename = NA)
}
\arguments{
\item{data}{data frame being used}

\item{date_var}{occurrence data variable being used}

\item{data_year}{year the data is for}

\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated results
}
\description{
Calculates Table 5.5 Deaths by place of occurrence and site of occurrence
}
\examples{
t5.5 <- create_t5_5(data = dth_data, date_var = dodyr, data_year = 2022, tablename = NA)

}
