% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_6_1.R
\name{create_t6.1}
\alias{create_t6.1}
\title{Create Table 6.1}
\usage{
create_t6.1(data, year_var, data_year)
}
\arguments{
\item{data}{data frame being used}

\item{data_year}{year of data}
}
\value{
Returns data frame with tabulated results
}
\description{
Create Table 6.1
}
\examples{
t6.1 <- create_t6.1(dth_data, dodyr, 2022)

}
