% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_6_2.R
\name{create_t6.2}
\alias{create_t6.2}
\title{Create Table 6.2}
\usage{
create_t6.2(data, date_var)
}
\arguments{
\item{data}{data frame being used}

\item{data_year}{year of data: dodyr or doryr}
}
\value{
Returns data frame with tabulated results
}
\description{
Create Table 6.2
}
\examples{
t6.2 <- create_t6.2(dth_data, dodyr)
}
