% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_7_6_to_7.9.R
\name{create_t7.6_to_7.9}
\alias{create_t7.6_to_7.9}
\title{Creates Tables 7.6 to 7.9}
\usage{
create_t7.6_to_7.9(
  data,
  data_year = 2019,
  group_var = "age_w",
  by_var = "age_h",
  tablename = NA
)
}
\arguments{
\item{data}{data frame being used}

\item{data_year}{year the data is for}

\item{group_var}{the tables columns}

\item{by_var}{tables rows}

\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated results
}
\description{
Creates Tables 7.6 to 7.9
}
\examples{
t7.6 <- create_t7.6_to_7.9(div_data, data_year = 2019, group_var = "age_w", by_var = "age_h", tablename = "Table_7_6")
t7.7 <- create_t7.6_to_7.9(div_data, data_year = 2019, group_var = "age_h", by_var = "dur_grp", tablename = "Table_7_7")
t7.8 <- create_t7.6_to_7.9(div_data, data_year = 2019, group_var = "age_w", by_var = "dur_grp", tablename = "Table_7_8")
t7.9 <- create_t7.6_to_7.9(div_data, data_year = 2019, group_var = "child", by_var = "dur_grp", tablename = "Table_7_9")


}
