% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_8_2.R
\name{create_t8.2}
\alias{create_t8.2}
\title{Creat Table 8.2}
\usage{
create_t8.2(data_year = 2022, tablename = NA)
}
\arguments{
\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated result
}
\description{
Creat Table 8.2
}
\examples{
t8.2 <- create_t8.2(tablename = "Table _8_2")

}
