% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_4_9_and _4_10.R
\name{create_table_4_9_and_4_10}
\alias{create_table_4_9_and_4_10}
\title{Creates Tables 4.9 and 4.10}
\usage{
create_table_4_9_and_4_10(
  data,
  est_data,
  data_year = 2022,
  ruindicator = "urban",
  tablename = NA
)
}
\arguments{
\item{data}{data frame being used}

\item{est_data}{data frame for estimate data}

\item{data_year}{year of data}

\item{ruindicator}{whether table is for urban or rural data}

\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated result
}
\description{
Creates Tables 4.9 and 4.10
}
\examples{
t4.9 <- create_table_4_9_and_4_10(data = bth_data, est_data = bth_est, data_year = 2022,
                                 ruindicator = "urban", tablename = "Table_4_9")
t4.10 <- create_table_4_9_and_4_10(data = bth_data, est_data = bth_est, data_year = 2022,
                                 ruindicator = "rural", tablename = "Table_4_10")
}
