% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_7_4_&_7_5.R
\name{create_table_7.4_and_7.5}
\alias{create_table_7.4_and_7.5}
\title{Create Tables 7.4 and 7.5}
\usage{
create_table_7.4_and_7.5(data, data_year, groombride = "groom", tablename = NA)
}
\arguments{
\item{data}{dataframe being used}

\item{data_year}{year the data is for}

\item{groombride}{whether the table is for groom or bride data}

\item{tablename}{name of the table being saved as a csv file}
}
\value{
data frame of tabulated results
}
\description{
Create Tables 7.4 and 7.5
}
\examples{
t7.4 <- create_table_7.4_and_7.5(marr_data, data_year = 2020, groombride = "groom", tablename = "Table_7_4")
t7.5 <- create_table_7.4_and_7.5(marr_data, data_year = 2020, groombride = "bride", tablename = "Table_7_5")

}
