% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_agegroup.R
\name{derive_age_groups}
\alias{derive_age_groups}
\title{derive_age_groups}
\usage{
derive_age_groups(
  age_in_yrs_col,
  start_age,
  max_band,
  step_size,
  under_1 = TRUE
)
}
\arguments{
\item{age_in_yrs_col}{name of the age in years variable used to derive group.}

\item{start_age}{this is the lowest age for the first group, excluding < 1,
if set to 5 then first group will be 01-04.}

\item{max_band}{this is the final band in the age groupings e.g. 85+.}

\item{step_size}{this is the width of each age band in years.}

\item{under_1}{set to TRUE if you require an 'under one year old' group.}
}
\value{
returns an agegroup based on set parameters
}
\description{
Derives agegroupings according to how the parameters are set.
}
\examples{
data <- data \%>\%
mutate(AGEGROUP = derive_age_groups(AGEINYRS, start_age = 5, max_band = 95,
step_size = 5, under_1 = TRUE))
data <- data \%>\%
mutate(AGEGROUP = derive_age_groups(AGEINYRS, start_age = 30, max_band = 95,
step_size = 5, under_1 = FALSE))

}
