% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_excel.R
\name{round_excel}
\alias{round_excel}
\title{round_excel}
\usage{
round_excel(x, n = 0)
}
\arguments{
\item{x}{Value or values to round}

\item{n}{Number of significant digits}
}
\value{
A rounded value or vector of values
}
\description{
A function to implement the Excel method of rounding.
}
\details{
Given a decimal, the function will round the value to the specified number of decimal
places using the Excel method of rounding (rounding up from .5).
}
\examples{
round_excel(c(1.6, 5.9, 2.5), 0)

}
