% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_4_1.R
\name{create_t4.1}
\alias{create_t4.1}
\title{Calculates Table 4.1 Births summary table}
\usage{
create_t4.1(data, tablename = "Table_4_1")
}
\arguments{
\item{data}{births data frame}

\item{tablename}{name for csv output use _ instead of . for names}
}
\value{
data frame with tabulated results
}
\description{
Calculates Table 4.1 Births summary table
}
\examples{
t4.1 <- create_t4.1(bth_data, tablename = "Table_4_1")

}
